pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "capiscio-sdk"
version = "0.1.0"
description = "Runtime security middleware for A2A agents"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [
    { name = "Capiscio Team", email = "team@capisc.io" }
]
keywords = ["a2a", "agent", "security", "validation", "middleware", "agent-to-agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Security",
]

dependencies = [
    "a2a-sdk>=0.1.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "cryptography>=42.0.0",
    "cachetools>=5.3.0",
    "pyjwt[crypto]>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
    "types-cachetools>=5.3.0",
]

[project.urls]
Homepage = "https://capisc.io"
Documentation = "https://docs.capisc.io/sdk-python"
Repository = "https://github.com/capiscio/capiscio-sdk-python"
Issues = "https://github.com/capiscio/capiscio-sdk-python/issues"

[tool.hatch.build.targets.wheel]
packages = ["capiscio_sdk"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312", "py313"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
